Refactor according to python names
