+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cityfront",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
