+++ template_settings.cfg
@@ -1,8 +1,13 @@
 [api]
 api_key = your_api_key
-model_path = ML/models/model_name.pkl
+model_path = ml/models/model_name.pkl
 
 [ml]
-raw_dataset = ML/data/raw/olx_house_price_Q122.csv
-model_dir = ML/models/
-custom_model = RandomForestRegressor(n_estimators=50, n_jobs=-1)
+raw_dataset = ml/data/raw/olx_house_price_Q122.csv
+model_dir = ml/models/
+custom_model = TransformedTargetRegressor(
+        regressor=RandomForestRegressor(n_estimators=50, n_jobs=-1),
+        func=np.log,
+        inverse_func=np.exp,
+        check_inverse=False,
+    )
